MODBULKOPS-398 - move tests for marc flow in marc package
