+++ src/main/java/org/folio/bulkops/processor/folio/FolioUpdateProcessorFactory.java
@@ -1,4 +1,4 @@
-package org.folio.bulkops.processor;
+package org.folio.bulkops.processor.folio;
 
 import java.util.List;
 import java.util.Map;
@@ -8,6 +8,7 @@ import org.folio.bulkops.domain.bean.ExtendedHoldingsRecord;
 import org.folio.bulkops.domain.bean.ExtendedInstance;
 import org.folio.bulkops.domain.bean.ExtendedItem;
 import org.folio.bulkops.domain.bean.User;
+import org.folio.bulkops.processor.FolioUpdateProcessor;
 import org.springframework.stereotype.Service;
 
 import jakarta.annotation.PostConstruct;
