+++ src/test/java/org/folio/bulkops/processor/marc/MarcInstanceDataProcessorTest.java
@@ -1,4 +1,4 @@
-package org.folio.bulkops.processor;
+package org.folio.bulkops.processor.marc;
 
 import static java.util.Objects.nonNull;
 import static org.assertj.core.api.Assertions.assertThat;
