+++ src/test/java/org/folio/bulkops/processor/marc/MarcInstanceUpdateProcessorTest.java
@@ -1,4 +1,4 @@
-package org.folio.bulkops.processor;
+package org.folio.bulkops.processor.marc;
 
 import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.assertj.core.api.Assertions.assertThat;
